fix file imports after folder rename
