+++ src/redux/reducers/dummyReducer.js
@@ -1,5 +1,5 @@
 // Import action creators needed for your specific reducer
-import { ADD_DUMMY } from "../actionCreators/actionTypes.js";
+import { ADD_DUMMY } from "../actions/actionTypes.js";
 
 // Example reducer
 function dummyReducer(state = ["First"], action) {
